Created first slice: AppSlice.
