+++ src/store/index.ts
@@ -1,11 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
-import reducer from './reducer';
 import { createAPI } from '../services/services';
+import rootReducer from './reducer';
 
 const api = createAPI();
 
 const store = configureStore({
-  reducer,
+  reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
