+++ src/types/store/selectors.ts
@@ -1,20 +1,21 @@
 import { AppState } from '.';
+import { NameSpace } from '../../components/consts';
 
-const selectOffers = (state: AppState) => state.offers;
-const selectComments = (state: AppState) => state.comments;
-const selectSortingType = (state: AppState) => state.sortingType;
-const selectCurrentCity = (state: AppState) => state.currentCity;
-const selectIsOfferDataLoading = (state: AppState) => state.isOfferDataLoading;
-const selectIsOffersDataLoading = (state: AppState) => state.isOffersDataLoading;
-const selectIsNearByDataLoading = (state: AppState) => state.isNearByDataLoading;
-const selectIsReviewsDataLoading = (state: AppState) => state.isReviewsDataLoading;
-const selectAuthorizationStatus = (state: AppState) => state.authorizationStatus;
-const selectCurrentUser = (state: AppState) => state.currentUser;
-const selectOffer = (state: AppState) => state.offer;
-const selectNearBys = (state: AppState) => state.nearBys;
-const selectReviews = (state: AppState) => state.reviews;
-const selectIsOffersError = (state: AppState) => state.isOffersError;
-const selectIsCommentsFetchingError = (state: AppState) => state.isCommentsFetchingError;
+const selectOffers = (state: AppState) => state[NameSpace.App].offers;
+const selectComments = (state: AppState) => state[NameSpace.App].comments;
+const selectSortingType = (state: AppState) => state[NameSpace.App].sortingType;
+const selectCurrentCity = (state: AppState): string => state[NameSpace.App].currentCity;
+const selectIsOfferDataLoading = (state: AppState) => state[NameSpace.App].isOfferDataLoading;
+const selectIsOffersDataLoading = (state: AppState) => state[NameSpace.App].isOffersDataLoading;
+const selectIsNearByDataLoading = (state: AppState) => state[NameSpace.App].isNearByDataLoading;
+const selectIsReviewsDataLoading = (state: AppState) => state[NameSpace.App].isReviewsDataLoading;
+const selectAuthorizationStatus = (state: AppState) => state[NameSpace.App].authorizationStatus;
+const selectCurrentUser = (state: AppState) => state[NameSpace.App].currentUser;
+const selectOffer = (state: AppState) => state[NameSpace.App].offer;
+const selectNearBys = (state: AppState) => state[NameSpace.App].nearBys;
+const selectReviews = (state: AppState) => state[NameSpace.App].reviews;
+const selectIsOffersError = (state: AppState) => state[NameSpace.App].isOffersError;
+const selectIsCommentsFetchingError = (state: AppState) => state[NameSpace.App].isCommentsFetchingError;
 
 export {
   selectOffers,
